Create init command on buildpack-packager
- Scaffold includes an entire buildpack of files, including out test
scripts and unimplented detect/compile/supply/finalize/release
- Using {{LANGUAGE}} as the place to template in the name of the buildpack
(e.g. python or nodejs or ruby)
- Using templates (text/template) to fill in LANGUAGE variables
- Copied a section of bindata.go into our file in order to apply template
to each file as it is created
- Assuming doesn't want -buildpack in the language name
- go generate will update bindata.go
- Make all the scripts & bin scripts executable in scaffold because
bindata.go will preserve permissions
- Make integration & brats test pending because app can't be pushed
without implementing release

Important milestone: Started using go-bindata

So, the original approach with files as in-line strings as a bad idea.
Instead, we're  moving to have the template directory live in this repo
as "scaffold". You can directly view and edit those files, and then use
go-bindata to regenerate the code representing them.

In order to include the current pattern of e.g. `/src/python`, we used
the directory name `LANGUAGE` and replace it with the buildpack name in
the code afterwards.

We are keeping the unit tests that check for file presence because that
is a failsafe against people forgetting/not knowing to run go-bindata.

[#155219487]