Implementing fetching dependencies from app cache dir.
- Add method for setting app cache dir on a manifest (needs to be
separate function because manifest is an argument to the Stager, but app
cache directory path is something that is set by the Stager).

- When an app cache dir is set by the buildpack on the Manifest (a
change in supply.go and sometimes finalize.go), in an uncached
buildpack, each dependency will be downloaded the first time it is needed and
copied from the cache every time after that.

- The buildpack is responsible for calling manifest.CleanUp() after it
is done installing all its dependencies. This will delete from the cache
all dependencies that were not just installed.

- Refactor the manifest tests to reuse code from the existing donwload &
cached buildpack tests for FetchDependency to also work for using the
app cache, with minimal duplication.

[#155405429]

Signed-off-by: Jack Feeny <jfeeny@pivotal.io>
Signed-off-by: Leah Hanson <lhanson@pivotal.io>