Make brats tests clean-up after themselves.
With the new stack-associated capi, if you create a buildpack again with
the same name, you create a nil-stack record.

When our brats tests 'CreateOrUpdate' a buildpack, they create a second
buildpack record now.

We've added a function to check how many copies of a given buildpack
there are, and updated the test to delete the buildpack twice.

In the future, we may want to make it use a smarter function than
'CreateOrUpdate' -- one that doesn't accidentally create a second
record.

[#158444917]

Signed-off-by: Jackson Feeny <jacksonfeeny@gmail.com>