Make brats improvement work on lts.
Because the lts environment does not have the new stack-associated capi,
the "expect two buildpacks" fails -- there will only be one.

So, we'll delete a second time if the first delete doesn't delete
everything. If we end up with more than two buildpacks eventually, the
test will fail -- which should prevent us from unexpectedly ending up
with stray buildpacks on edge in the future.

[#158444917]

Signed-off-by: Jackson Feeny <jacksonfeeny@gmail.com>