Don't try to match the dependency file name
The test is only trying to make sure that at least one dependency was
installed (and the credentials were redacted) and that the credentials
are nowhere to be found in the output.

This change achieves that without the need to match a specific
dependency filename. This makes it possible to test buildpacks with
filenames that don't match the upstream naming conventions.